chore(BUX-164): move CMP into separate file
